+++ bin/cyclonedx-npm-cli.js
@@ -1,2 +1,7 @@
 #!/usr/bin/env node
-require('../dist/cli.js').run(process)
+try {
+  require('../dist/cli.js').run(process)
+} catch (e) {
+  process.stderr.write(`\n${e}\n`)
+  process.exit(e?.code || 1)
+}