+++ src/cli.ts
@@ -49,7 +49,7 @@ interface CommandOptions {
   mcType: Enums.ComponentType
 }
 
-function makeCommand (): Command {
+function makeCommand (process: NodeJS.Process): Command {
   return new Command(
   ).description(
     'Create CycloneDX Software Bill of Materials (SBOM) from Node.js NPM projects.'
@@ -161,7 +161,7 @@ export function run (process: NodeJS.Process): void {
   // all output shall be bound to stdError - stdOut is for result output only
   const myConsole = new console.Console(process.stderr, process.stderr)
 
-  const program = makeCommand()
+  const program = makeCommand(process)
   program.parse(process.argv)
 
   const options: CommandOptions = program.opts()
@@ -169,9 +169,7 @@ export function run (process: NodeJS.Process): void {
 
   const packageFile = resolve(process.cwd(), program.args[0] ?? 'package.json')
   if (!existsSync(packageFile)) {
-    const msg = `missing project's manifest file: ${packageFile}`
-    program.error(msg)
-    throw new Error(msg)
+    throw new Error(`missing project's manifest file: ${packageFile}`)
   }
   myConsole.debug('DEBUG | packageFile: %s', packageFile)
   const projectDir = dirname(packageFile)
@@ -192,9 +190,7 @@ export function run (process: NodeJS.Process): void {
   } else if (existsSync(packageLockFile)) {
     lockFile = packageLockFile
   } else {
-    const msg = 'missing package lock file, missing npm shrinkwrap file'
-    program.error(msg)
-    throw new Error(msg)
+    throw new Error('missing package lock file or npm shrinkwrap file')
   }
   myConsole.debug('DEBUG | lockFile: %s', lockFile)
 
@@ -216,13 +212,11 @@ export function run (process: NodeJS.Process): void {
       flattenComponents: options.flattenComponents
     },
     myConsole
-  ).buildFromLockFile(lockFile)
+  ).buildFromLockFile(lockFile, process)
 
   const spec = Spec.SpecVersionDict[options.specVersion]
   if (undefined === spec) {
-    const msg = 'unsupported spec-version'
-    program.error(msg)
-    throw new Error(msg)
+    throw new Error('unsupported spec-version')
   }
 
   let serializer: Serialize.Types.Serializer