+++ HISTORY.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## unreleased
 
+* Build
+  * Use _TypeScript_ `v4.8.2` now, was `v4.7.4`. (via [#78])
+
+[#78]: https://github.com/CycloneDX/cyclonedx-javascript-library/pull/78
+
 ## 1.0.0-beta.4
 
 * Fixed