+++ HISTORY.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## unreleased
 
+## 1.0.0-beta.5
+
 * Added
   * Support for node 14 was enabled.
   * Support for handling when run via `npx`.